+++ packages/documentation/public/how-to-contribute/about.md
@@ -1,0 +1,3 @@
+In this section you will find all the information you need before starting your contribution to Offcourse. This includes the contributing guidelines and the work ethos we aspire as a community.
+
+If anything is unclear, you can have a look at the [FAQ](https://www.offcourse.io/faq) on the Offcourse website, or contact the maintainers of this website via [contact@offcourse.io](contact@offcourse.io).
