--- packages/documentation/public/product-roadmap/intro.md
@@ -1,8 +1,0 @@
-# Product roadmap
-
-Curious to know what features you can help us develop? 
-
-You will see on our roadmap that there are a lot of opportunities to expand the functionalities of the platform, both the functionality of the course cards as well as navigation on the platform for example. 
-We continuously update our product roadmap with new features. 
-
-We are also open to discuss features you would like to work on that are not on our roadmap. You can get in touch via [email](mailto:contact@offcourse.io) to discuss the possibilities.
