+++ packages/documentation/src/App.js
@@ -1,29 +1,7 @@
 import React, { Component } from "react";
-import * as atoms from "@offcourse/atoms";
-import * as molecules from "@offcourse/molecules";
-import * as organisms from "@offcourse/organisms";
-import { Catalog, pageLoader } from "catalog";
+import pages from "./pages";
+import { Catalog } from "catalog";
 import { offcourse } from "./themes";
-
-const baseUrl = `https://github.com/OffCourse/offcourse-next/tree/master/packages/atoms/src/`;
-
-const createPages = ({ name: collectionName, blocks, helpers }) => {
-  const pageNames = Object.keys(blocks);
-  return pageNames.map(blockName => ({
-    imports: {
-      ...blocks,
-      ...helpers,
-      meta: {
-        blockName,
-        collectionName,
-        url: baseUrl + blockName
-      }
-    },
-    path: `/${collectionName}/${blockName}`,
-    title: `${blockName}`,
-    content: pageLoader(`/${collectionName}/${blockName}.md`)
-  }));
-};
 
 const { fonts, grayScale } = offcourse;
 
@@ -41,43 +19,7 @@ class App extends Component {
         title="Documentation"
         theme={catalogTheme}
         useBrowserHistory={true}
-        pages={[
-          {
-            path: "/",
-            title: "General Introduction",
-            content: pageLoader("./general-introduction.md")
-          },
-          {
-            path: "/visual-identity",
-            title: "Visual Identity",
-            imports: {
-              ...offcourse.colors,
-              grayScale: offcourse.namedGrayScale
-            },
-            content: pageLoader("./visual-identity.md")
-          },
-          {
-            title: "Atoms",
-            pages: createPages({
-              name: "atoms",
-              blocks: atoms
-            })
-          },
-          {
-            title: "Molecules",
-            pages: createPages({
-              name: "molecules",
-              blocks: molecules
-            })
-          },
-          {
-            title: "Organisms",
-            pages: createPages({
-              name: "organisms",
-              blocks: organisms
-            })
-          }
-        ]}
+        pages={pages}
       />
     );
   }
